requirements.txt
requests
pandas
streamlit
openai
google-generativeai
python-dotenv
google-cloud-bigquery